chore: improve Heading exports
